# docs/openapi/swagger.json
{
  "openapi" : "3.0.1",
  "info" : {
    "contact" : {
      "email" : "support@swagger.io",
      "name" : "API Support",
      "url" : "http://www.swagger.io/support"
    },
    "description" : "LocalPaas Backend",
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "termsOfService" : "http://swagger.io/terms/",
    "title" : "LocalPaas Backend",
    "version" : "0.1"
  },
  "servers" : [ {
    "url" : "/v1"
  } ],
  "paths" : {
...
    "/cluster/nodes" : {
      "get" : {
        "description" : "Lists cluster nodes",
        "operationId" : "listNode",
        "parameters" : [ {
          "description" : "`status=<target>`",
          "in" : "query",
          "name" : "status",
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "`search=<target> (support *)`",
          "in" : "query",
          "name" : "search",
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "`pageOffset=offset`",
          "in" : "query",
          "name" : "pageOffset",
          "schema" : {
            "type" : "integer"
          }
        }, {
          "description" : "`pageLimit=limit`",
          "in" : "query",
          "name" : "pageLimit",
          "schema" : {
            "type" : "integer"
          }
        }, {
          "description" : "`sort=[-]field1|field2...`",
          "in" : "query",
          "name" : "sort",
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_clusteruc_clusterdto.ListNodeResp"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo"
                }
              }
            },
            "description" : "Bad Request"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Lists cluster nodes",
        "tags" : [ "cluster_nodes" ]
      }
    },
    "/cluster/nodes/{nodeID}" : {
      "delete" : {
        "description" : "Deletes a node",
        "operationId" : "deleteNode",
        "parameters" : [ {
          "description" : "node ID",
          "in" : "path",
          "name" : "nodeID",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_clusteruc_clusterdto.DeleteNodeResp"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo"
                }
              }
            },
            "description" : "Bad Request"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Deletes a node",
        "tags" : [ "cluster_nodes" ]
      },
      "get" : {
        "description" : "Gets node details",
        "operationId" : "getNode",
        "parameters" : [ {
          "description" : "node ID",
          "in" : "path",
          "name" : "nodeID",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_clusteruc_clusterdto.GetNodeResp"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo"
                }
              }
            },
            "description" : "Bad Request"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Gets node details",
        "tags" : [ "cluster_nodes" ]
      }
    },
...
    "/projects" : {
      "get" : {
        "description" : "Lists projects",
        "operationId" : "listProject",
        "parameters" : [ {
          "description" : "`status=<target>`",
          "in" : "query",
          "name" : "status",
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "`search=<target> (support *)`",
          "in" : "query",
          "name" : "search",
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "`pageOffset=offset`",
          "in" : "query",
          "name" : "pageOffset",
          "schema" : {
            "type" : "integer"
          }
        }, {
          "description" : "`pageLimit=limit`",
          "in" : "query",
          "name" : "pageLimit",
          "schema" : {
            "type" : "integer"
          }
        }, {
          "description" : "`sort=[-]field1|field2...`",
          "in" : "query",
          "name" : "sort",
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_projectuc_projectdto.ListProjectResp"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo"
                }
              }
            },
            "description" : "Bad Request"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Lists projects",
        "tags" : [ "projects" ]
      },
      "post" : {
        "description" : "Creates a new project",
        "operationId" : "createProject",
        "requestBody" : {
          "content" : {
            "*/*" : {
              "schema" : {
                "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_projectuc_projectdto.CreateProjectReq"
              }
            }
          },
          "description" : "request data",
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_projectuc_projectdto.CreateProjectResp"
                }
              }
            },
            "description" : "Created"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo"
                }
              }
            },
            "description" : "Bad Request"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Creates a new project",
        "tags" : [ "projects" ],
        "x-codegen-request-body-name" : "body"
      }
    },
...
    "/projects/{projectID}" : {
      "delete" : {
        "description" : "Deletes a project",
        "operationId" : "deleteProject",
        "parameters" : [ {
          "description" : "project ID",
          "in" : "path",
          "name" : "projectID",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_projectuc_projectdto.DeleteProjectResp"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo"
                }
              }
            },
            "description" : "Bad Request"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Deletes a project",
        "tags" : [ "projects" ]
      },
      "get" : {
        "description" : "Gets project details",
        "operationId" : "getProject",
        "parameters" : [ {
          "description" : "project ID",
          "in" : "path",
          "name" : "projectID",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_projectuc_projectdto.GetProjectResp"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo"
                }
              }
            },
            "description" : "Bad Request"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Gets project details",
        "tags" : [ "projects" ]
      }
    },
...
    "/sessions/me" : {
      "get" : {
        "description" : "Gets session info of the current user",
        "operationId" : "getMe",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_sessionuc_sessiondto.GetMeResp"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo"
                }
              }
            },
            "description" : "Bad Request"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Gets session info of the current user",
        "tags" : [ "sessions" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo" : {
        "properties" : {
          "cause" : {
            "type" : "string"
          },
          "code" : {
            "type" : "string"
          },
          "debugLog" : {
            "type" : "string"
          },
          "detail" : {
            "type" : "string"
          },
          "errors" : {
            "items" : {
              "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.InnerErrorInfo"
            },
            "type" : "array"
          },
          "stackTrace" : {
            "type" : "string"
          },
          "status" : {
            "type" : "integer"
          },
          "title" : {
            "type" : "string"
          },
          "type" : {
            "type" : "string"
          }
        },
        "required" : [ "code", "detail", "status", "title", "type" ],
        "type" : "object"
      },
      "github_com_localpaas_localpaas_localpaas_app_apperrors.InnerErrorInfo" : {
        "properties" : {
          "cause" : {
            "type" : "string"
          },
          "code" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          },
          "path" : {
            "type" : "string"
          }
        },
        "required" : [ "code", "message", "path" ],
        "type" : "object"
      },
...
      "github_com_localpaas_localpaas_localpaas_app_base.NodeRole" : {
        "enum" : [ "manager", "worker" ],
        "type" : "string",
        "x-enum-varnames" : [ "NodeRoleManager", "NodeRoleWorker" ]
      },
      "github_com_localpaas_localpaas_localpaas_app_base.NodeStatus" : {
        "enum" : [ "unknown", "down", "ready", "disconnected" ],
        "type" : "string",
        "x-enum-varnames" : [ "NodeStatusUnknown", "NodeStatusDown", "NodeStatusReady", "NodeStatusDisconnected" ]
      },
      "github_com_localpaas_localpaas_localpaas_app_base.ProjectStatus" : {
        "enum" : [ "active", "locked", "disabled", "deleting" ],
        "type" : "string",
        "x-enum-varnames" : [ "ProjectStatusActive", "ProjectStatusLocked", "ProjectStatusDisabled", "ProjectStatusDeleting" ]
      },
...
      "github_com_localpaas_localpaas_localpaas_app_basedto.BaseMeta" : {
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "github_com_localpaas_localpaas_localpaas_app_basedto.Meta" : {
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          },
          "page" : {
            "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_basedto.PagingMeta"
          }
        },
        "type" : "object"
      },
...
      "github_com_localpaas_localpaas_localpaas_app_basedto.ObjectIDResp" : {
        "properties" : {
          "id" : {
            "type" : "string"
          }
        },
        "required" : [ "id" ],
        "type" : "object"
      },
      "github_com_localpaas_localpaas_localpaas_app_basedto.PagingMeta" : {
        "properties" : {
          "limit" : {
            "type" : "integer"
          },
          "offset" : {
            "type" : "integer"
          },
          "total" : {
            "type" : "integer"
          }
        },
        "required" : [ "limit", "offset", "total" ],
        "type" : "object"
      },
      "github_com_localpaas_localpaas_localpaas_app_entity.AccessActions" : {
        "properties" : {
          "delete" : {
            "type" : "boolean"
          },
          "read" : {
            "type" : "boolean"
          },
          "write" : {
            "type" : "boolean"
          }
        },
        "required" : [ "delete", "read", "write" ],
        "type" : "object"
      },
...
      "github_com_localpaas_localpaas_localpaas_app_usecase_clusteruc_clusterdto.DeleteNodeResp" : {
        "properties" : {
          "meta" : {
            "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_basedto.BaseMeta"
          }
        },
        "required" : [ "meta" ],
        "type" : "object"
      },
      "github_com_localpaas_localpaas_localpaas_app_usecase_clusteruc_clusterdto.GetNodeResp" : {
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_clusteruc_clusterdto.NodeResp"
          },
          "meta" : {
            "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_basedto.BaseMeta"
          }
        },
        "required" : [ "data", "meta" ],
        "type" : "object"
      },
      "github_com_localpaas_localpaas_localpaas_app_usecase_clusteruc_clusterdto.ListNodeResp" : {
        "properties" : {
          "data" : {
            "items" : {
              "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_clusteruc_clusterdto.NodeResp"
            },
            "type" : "array"
          },
          "meta" : {
            "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_basedto.Meta"
          }
        },
        "required" : [ "data", "meta" ],
        "type" : "object"
      },
      "github_com_localpaas_localpaas_localpaas_app_usecase_clusteruc_clusterdto.NodePlatformResp" : {
        "properties" : {
          "architecture" : {
            "type" : "string"
          },
          "os" : {
            "type" : "string"
          }
        },
        "required" : [ "architecture", "os" ],
        "type" : "object"
      },
      "github_com_localpaas_localpaas_localpaas_app_usecase_clusteruc_clusterdto.NodeResources" : {
        "properties" : {
          "memoryBytes" : {
            "type" : "integer"
          },
          "nanoCPUs" : {
            "type" : "integer"
          }
        },
        "required" : [ "memoryBytes", "nanoCPUs" ],
        "type" : "object"
      },
      "github_com_localpaas_localpaas_localpaas_app_usecase_clusteruc_clusterdto.NodeResp" : {
        "properties" : {
          "addr" : {
            "type" : "string"
          },
          "createdAt" : {
            "type" : "string"
          },
          "hostname" : {
            "type" : "string"
          },
          "id" : {
            "type" : "string"
          },
          "isLeader" : {
            "type" : "boolean"
          },
          "platform" : {
            "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_clusteruc_clusterdto.NodePlatformResp"
          },
          "resources" : {
            "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_clusteruc_clusterdto.NodeResources"
          },
          "role" : {
            "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_base.NodeRole"
          },
          "status" : {
            "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_base.NodeStatus"
          },
          "updatedAt" : {
            "type" : "string"
          }
        },
        "required" : [ "addr", "createdAt", "hostname", "id", "isLeader", "platform", "resources", "role", "status", "updatedAt" ],
        "type" : "object"
      },
      "github_com_localpaas_localpaas_localpaas_app_usecase_projectuc_projectdto.CreateProjectReq" : {
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "note" : {
            "type" : "string"
          },
          "status" : {
            "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_base.ProjectStatus"
          },
          "tags" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          }
        },
        "required" : [ "name", "note", "status", "tags" ],
        "type" : "object"
      },
      "github_com_localpaas_localpaas_localpaas_app_usecase_projectuc_projectdto.CreateProjectResp" : {
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_basedto.ObjectIDResp"
          },
          "meta" : {
            "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_basedto.BaseMeta"
          }
        },
        "required" : [ "data", "meta" ],
        "type" : "object"
      },
...
      "github_com_localpaas_localpaas_localpaas_app_usecase_projectuc_projectdto.DeleteProjectResp" : {
        "properties" : {
          "meta" : {
            "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_basedto.BaseMeta"
          }
        },
        "required" : [ "meta" ],
        "type" : "object"
      },
...
      "github_com_localpaas_localpaas_localpaas_app_usecase_projectuc_projectdto.GetProjectResp" : {
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_projectuc_projectdto.ProjectResp"
          },
          "meta" : {
            "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_basedto.BaseMeta"
          }
        },
        "required" : [ "data", "meta" ],
        "type" : "object"
      },
...
      "github_com_localpaas_localpaas_localpaas_app_usecase_projectuc_projectdto.ListProjectResp" : {
        "properties" : {
          "data" : {
            "items" : {
              "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_projectuc_projectdto.ProjectResp"
            },
            "type" : "array"
          },
          "meta" : {
            "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_basedto.Meta"
          }
        },
        "required" : [ "data", "meta" ],
        "type" : "object"
      },
      "github_com_localpaas_localpaas_localpaas_app_usecase_projectuc_projectdto.ProjectAppResp" : {
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          }
        },
        "required" : [ "id", "name" ],
        "type" : "object"
      },
...
      "github_com_localpaas_localpaas_localpaas_app_usecase_projectuc_projectdto.ProjectResp" : {
        "properties" : {
          "apps" : {
            "items" : {
              "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_projectuc_projectdto.ProjectAppResp"
            },
            "type" : "array"
          },
          "createdAt" : {
            "type" : "string"
          },
          "id" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "note" : {
            "type" : "string"
          },
          "photo" : {
            "type" : "string"
          },
          "slug" : {
            "type" : "string"
          },
          "status" : {
            "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_base.ProjectStatus"
          },
          "tags" : {
            "description" : "manual copy ProjectTag -> string",
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "updatedAt" : {
            "type" : "string"
          },
          "userAccesses" : {
            "items" : {
              "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_projectuc_projectdto.ProjectUserAccessResp"
            },
            "type" : "array"
          }
        },
        "required" : [ "apps", "createdAt", "id", "name", "note", "photo", "slug", "status", "tags", "updatedAt", "userAccesses" ],
        "type" : "object"
      },
...
      "github_com_localpaas_localpaas_localpaas_app_usecase_projectuc_projectdto.ProjectUserAccessResp" : {
        "properties" : {
          "access" : {
            "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_entity.AccessActions"
          },
          "fullName" : {
            "type" : "string"
          },
          "id" : {
            "type" : "string"
          },
          "photo" : {
            "type" : "string"
          }
        },
        "required" : [ "access", "fullName", "id", "photo" ],
        "type" : "object"
      },
...
      "github_com_localpaas_localpaas_localpaas_app_usecase_sessionuc_sessiondto.GetMeDataResp" : {
        "properties" : {
          "user" : {
            "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_sessionuc_sessiondto.UserResp"
          }
        },
        "required" : [ "user" ],
        "type" : "object"
      },
      "github_com_localpaas_localpaas_localpaas_app_usecase_sessionuc_sessiondto.GetMeResp" : {
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_sessionuc_sessiondto.GetMeDataResp"
          },
          "meta" : {
            "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_basedto.BaseMeta"
          }
        },
        "required" : [ "data", "meta" ],
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "BasicAuth" : {
        "scheme" : "basic",
        "type" : "http"
      }
    }
  },
...
}